Hack to fix problem when connection pool fills up.
When the connection pool fills up, and _mexe tries to return a
connection to the pool, put_http_connection was closing the
connection.  This was a problem because the caller has not yet
read the body of the response.

The hack-y fix is to not close the connection, and delay the
closing until the response object is freed.